Fix root tag serialization
According to the NBT specification, the root Tag of an NBT document
is implicitly Compound, and the root tag type is not included.
However, it's theoretically possible to write out the tag -- thereby
allowing a value of any type to be the root node.  Therefore, an
option is presented to do so.